+++ wandb/errors/__init__.py
@@ -11,7 +11,6 @@ __all__ = [
     "LaunchError",
     "SweepError",
     "WaitTimeoutError",
-    "ContextCancelledError",
 ]
 
 from typing import List, Optional
@@ -123,17 +122,5 @@ class SweepError(Error):
 
 class WaitTimeoutError(Error):
     """Raised when wait() timeout occurs before process is finished"""
-
-    pass
-
-
-class MailboxError(Error):
-    """Generic Mailbox Exception"""
-
-    pass
-
-
-class ContextCancelledError(Error):
-    """Context cancelled Exception"""
 
     pass
