+++ wandb/sdk/internal/sender.py
@@ -29,10 +29,11 @@ import requests
 
 import wandb
 from wandb import util
-from wandb.errors import CommError, ContextCancelledError
+from wandb.errors import CommError
 from wandb.filesync.dir_watcher import DirWatcher
 from wandb.proto import wandb_internal_pb2
 from wandb.sdk.lib import redirect
+from wandb.sdk.lib.mailbox import ContextCancelledError
 
 from ..interface import interface
 from ..interface.interface_queue import InterfaceQueue
