+++ wandb/sdk/wandb_run.py
@@ -39,7 +39,6 @@ from wandb._globals import _datatypes_set_callback
 from wandb.apis import internal, public
 from wandb.apis.internal import Api
 from wandb.apis.public import Api as PublicApi
-from wandb.errors import MailboxError
 from wandb.proto.wandb_internal_pb2 import (
     MetricRecord,
     PollExitResponse,
@@ -82,7 +81,7 @@ from .lib import (
 )
 from .lib.exit_hooks import ExitHooks
 from .lib.git import GitRepo
-from .lib.mailbox import MailboxHandle, MailboxProbe, MailboxProgress
+from .lib.mailbox import MailboxError, MailboxHandle, MailboxProbe, MailboxProgress
 from .lib.printer import get_printer
 from .lib.proto_util import message_to_dict
 from .lib.reporting import Reporter
